backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
google-cloud-vision==3.4.5
google-generativeai==0.3.1
pdf2image==1.16.3
Pillow==10.1.0
python-dotenv==1.0.0
